+++ example/python/files/test_sample.py
@@ -5,6 +5,8 @@ def func(number):
     """test function"""
     if number is 4:
         return number + 1
+    elif number is 5:
+        return number + 1
     else:
         return number