+++ src/indicators.ts
@@ -1,9 +1,11 @@
 import {IConfigStore} from "./config";
+import {InterfaceXmlParse} from "./wrappers/xml-parse";
 import {InterfaceLcovParse} from "./wrappers/lcov-parse";
 import {InterfaceVscode} from "./wrappers/vscode";
 
-import {LcovSection} from "lcov-parse";
+import {Section} from "lcov-parse";
 import {Range, TextEditor} from "vscode";
+import {extname} from "path";
 
 export interface ICoverageLines {
     full: Range[];
@@ -12,21 +14,24 @@ export interface ICoverageLines {
 }
 
 export class Indicators {
-    private parse: InterfaceLcovParse;
+    private lcovParse: InterfaceLcovParse;
+    private xmlParse: InterfaceXmlParse;
     private vscode: InterfaceVscode;
     private configStore: IConfigStore;
 
     constructor(
-        parse: InterfaceLcovParse,
+        xmlParse: InterfaceXmlParse,
+        lcovParse: InterfaceLcovParse,
         vscode: InterfaceVscode,
         configStore: IConfigStore,
     ) {
-        this.parse = parse;
+        this.lcovParse = lcovParse;
+        this.xmlParse = xmlParse;
         this.vscode = vscode;
         this.configStore = configStore;
     }
 
-    public renderToTextEditor(section: LcovSection, textEditor: TextEditor): Promise<string> {
+    public renderToTextEditor(section: Section, textEditor: TextEditor): Promise<string> {
         return new Promise<string>((resolve, reject) => {
             const coverageLines: ICoverageLines = {
                 full: [],
@@ -41,10 +46,35 @@ export class Indicators {
         });
     }
 
-    public extract(lcovFile: string, file: string): Promise<LcovSection> {
-        if (lcovFile === "") { return Promise.reject("No coverage details inside lcov file!"); }
-        return new Promise<LcovSection>((resolve, reject) => {
-            this.parse.source(lcovFile, (err, data) => {
+    public async extractCoverage(coverageFile: string, file: string): Promise<Section> {
+        if(coverageFile.includes("<?xml")) {
+            return this.xmlExtract(coverageFile, file);
+        } else {
+            return this.lcovExtract(coverageFile, file);
+        }
+    }
+
+    private xmlExtract(xmlFile: string, file: string): Promise<Section> {
+        if (xmlFile === "") { return Promise.reject("No coverage details inside file!"); }
+        return new Promise<Section>((resolve, reject) => {
+            this.xmlParse.parseContent(xmlFile, (err, data) => {
+                if (err) { return reject(err); }
+                const section = data.find((lcovSection) => {
+                    const rootFolder = this.vscode.getRootPath().split(/[\\\/]/).reverse()[0];
+                    const coverageFile = `${rootFolder}/${lcovSection.file}`
+                    return this.compareFilePaths(coverageFile, file);
+                });
+
+                if (!section) { return reject(new Error("No coverage for file!")); }
+                return resolve(section);
+            });
+        });
+    }
+
+    private lcovExtract(lcovFile: string, file: string): Promise<Section> {
+        if (lcovFile === "") { return Promise.reject("No coverage details inside file!"); }
+        return new Promise<Section>((resolve, reject) => {
+            this.lcovParse.source(lcovFile, (err, data) => {
                 if (err) { return reject(err); }
                 const section = data.find((lcovSection) => {
                     return this.compareFilePaths(lcovSection.file, file);
@@ -67,7 +97,7 @@ export class Indicators {
         editor.setDecorations(this.configStore.partialCoverageDecorationType, coverage.partial);
     }
 
-    private filterCoverage(section: LcovSection, coverageLines: ICoverageLines): ICoverageLines {
+    private filterCoverage(section: Section, coverageLines: ICoverageLines): ICoverageLines {
         section.lines.details.forEach((detail) => {
             const lineRange = new Range(detail.line - 1, 0, detail.line - 1, 0);
             if (detail.hit > 0) {
@@ -77,23 +107,25 @@ export class Indicators {
             }
         });
 
-        section.branches.details.forEach((detail) => {
-            if (detail.branch === 0 && detail.taken === 0) {
-                const partialRange = new Range(detail.line - 1, 0, detail.line - 1, 0);
-                if (coverageLines.full.find((range) => range.isEqual(partialRange))) {
-                    // remove full converage if partial is a better match
-                    coverageLines.full = coverageLines.full.filter((range) => !range.isEqual(partialRange));
-                    coverageLines.partial.push(partialRange);
+        if (section.branches) {
+            section.branches.details.forEach((detail) => {
+                if (detail.branch === 0 && detail.taken === 0) {
+                    const partialRange = new Range(detail.line - 1, 0, detail.line - 1, 0);
+                    if (coverageLines.full.find((range) => range.isEqual(partialRange))) {
+                        // remove full converage if partial is a better match
+                        coverageLines.full = coverageLines.full.filter((range) => !range.isEqual(partialRange));
+                        coverageLines.partial.push(partialRange);
+                    }
                 }
-            }
-        });
+            });
+        }
         return coverageLines;
     }
 
-    private compareFilePaths(lcovFile: string, file: string): boolean {
+    private compareFilePaths(coverageFile: string, file: string): boolean {
         if (this.configStore.altSfCompare) {
             // consider windows and linux file paths
-            const sourceFile = lcovFile.split(/[\\\/]/).reverse();
+            const sourceFile = coverageFile.split(/[\\\/]/).reverse();
             const openFile = file.split(/[\\\/]/).reverse();
             const folderName = this.vscode.getRootPath().split(/[\\\/]/).reverse()[0];
             let match = true;
@@ -112,7 +144,7 @@ export class Indicators {
             return match;
         } else {
             // prevent hazardous casing mishaps
-            return lcovFile.toLocaleLowerCase() === file.toLocaleLowerCase();
+            return coverageFile.toLocaleLowerCase() === file.toLocaleLowerCase();
         }
     }
 }